move or replace old general ceremony code
