+++ scripts/events.py
@@ -14,7 +14,6 @@ from scripts.config import get_config
 
 # pylint: enable=line-too-long
 import traceback
-from math import floor
 
 import i18n
 
@@ -28,7 +27,6 @@ from scripts.cat.enums import (
 )
 from scripts.cat.names import Name
 from scripts.cat.save_load import save_cats, add_cat_to_fade_id
-from scripts.cat.skills import SkillPath
 from scripts.clan_package.settings import get_clan_setting, set_clan_setting
 from scripts.clan_resources.freshkill import FRESHKILL_EVENT_ACTIVE
 from scripts.conditions import (
@@ -37,6 +35,10 @@ from scripts.conditions import (
 )
 from scripts.cat.microservices.conditions import get_ill, get_injured
 from scripts.event_class import Single_Event
+from scripts.events_module.ceremony.perform_ceremony import (
+    perform_ceremonies,
+    trigger_ceremony,
+)
 
 from scripts.events_module.generate_events import GenerateEvents, generate_events
 from scripts.events_module.outsider import outsider_events
@@ -58,7 +60,6 @@ from scripts.ui.windows.save_error import SaveErrorWindow
 from scripts.events_module.text_adjust import (
     ongoing_event_text_adjust,
     event_text_adjust,
-    ceremony_text_adjust,
     adjust_list_text,
     history_text_adjust,
 )
@@ -79,11 +80,8 @@ logger = logging.getLogger(__name__)
 all_events = {}
 new_cat_invited = False
 ceremony_accessory = False
-CEREMONY_TXT = None
 WAR_TXT = None
-ceremony_lang = None
 war_lang = None
-ceremony_id_by_tag = {}
 
 
 def one_moon():
@@ -871,13 +869,13 @@ def handle_lost_cats_return(predetermined_cat_IDs: list = None):
         ]:
             if x.moons >= 15:
                 if x.status.rank == CatRank.MEDICINE_APPRENTICE:
-                    ceremony(x, CatRank.MEDICINE_CAT)
+                    trigger_ceremony(x, CatRank.MEDICINE_CAT)
                 elif x.status.rank == CatRank.MEDIATOR_APPRENTICE:
-                    ceremony(x, CatRank.MEDIATOR)
+                    trigger_ceremony(x, CatRank.MEDIATOR)
                 else:
-                    ceremony(x, CatRank.WARRIOR)
+                    trigger_ceremony(x, CatRank.WARRIOR)
             elif not x.status.rank.is_any_apprentice_rank() and x.moons >= 6:
-                ceremony(x, CatRank.APPRENTICE)
+                trigger_ceremony(x, CatRank.APPRENTICE)
 
 
 def handle_fading(cat):
@@ -1224,596 +1222,6 @@ def check_war():
         clan=game.clan,
     )
     game.cur_events_list.append(Single_Event(event, "other_clans"))
-
-
-def perform_ceremonies(cat):
-    """
-    ceremonies
-    """
-
-    global ceremony_accessory
-
-    # Protection check, to ensure "None" cats won't cause a crash.
-    if not cat or cat.dead:
-        return
-
-    if cat.status.rank == CatRank.DEPUTY and game.clan.deputy is None:
-        game.clan.deputy = cat
-    if cat.status.rank == CatRank.MEDICINE_CAT and game.clan.medicine_cat is None:
-        game.clan.medicine_cat = cat
-
-    # PROMOTE DEPUTY TO LEADER, IF NEEDED -----------------------
-
-    # If a Clan deputy exists, and the leader is dead,
-    #  outside, or doesn't exist, make the deputy leader.
-    if cat == game.clan.deputy:
-        # leader gone, time to promote
-        if not game.clan.leader or not game.clan.leader.status.alive_in_player_clan:
-            game.clan.leader_lives = 9
-            ceremony(cat, CatRank.LEADER)
-            cat.generate_lead_ceremony()
-            game.clan.deputy = None
-            game.clan.leader = cat
-
-    # OTHER CEREMONIES ---------------------------------------
-
-    # retiring to elder den
-    if (
-        not cat.no_retire
-        and cat.status.rank in (CatRank.WARRIOR, CatRank.DEPUTY)
-        and len(cat.apprentice) < 1
-        and cat.moons > 114
-    ):
-        # There is some variation in the age.
-        if cat.moons > 140 or not int(random.random() * (-0.7 * cat.moons + 100)):
-            if cat.status.rank == CatRank.DEPUTY:
-                game.clan.deputy = None
-            ceremony(cat, CatRank.ELDER)
-
-    # apprentice a kitten to either med or warrior
-    if cat.moons == cat_class.age_moons[CatAge.ADOLESCENT][0]:
-        if cat.status.rank == CatRank.KITTEN:
-            if _is_suitable_medcat_app(cat):
-                ceremony(cat, CatRank.MEDICINE_APPRENTICE)
-                ceremony_accessory = True
-                gain_accessories(cat)
-            else:
-                # Chance for mediator apprentice
-                mediator_list = list(
-                    filter(
-                        lambda x: x.status.rank == CatRank.MEDIATOR
-                        and x.status.alive_in_player_clan,
-                        Cat.all_cats_list,
-                    )
-                )
-
-                # This checks if at least one mediator already has an apprentice.
-                has_mediator_apprentice = False
-                for c in mediator_list:
-                    if c.apprentice:
-                        has_mediator_apprentice = True
-                        break
-
-                chance = constants.CONFIG["roles"]["mediator_app_chance"]
-                if cat.personality.trait in [
-                    "charismatic",
-                    "loving",
-                    "responsible",
-                    "wise",
-                    "thoughtful",
-                ]:
-                    chance = int(chance / 1.5)
-                if cat.is_disabled():
-                    chance = int(chance / 2)
-
-                if chance == 0:
-                    chance = 1
-
-                # Only become a mediator if there is already one in the clan.
-                if (
-                    mediator_list
-                    and not has_mediator_apprentice
-                    and not int(random.random() * chance)
-                ):
-                    ceremony(cat, CatRank.MEDIATOR_APPRENTICE)
-                    ceremony_accessory = True
-                    gain_accessories(cat)
-                else:
-                    ceremony(cat, CatRank.APPRENTICE)
-                    ceremony_accessory = True
-                    gain_accessories(cat)
-
-    # graduate
-    if cat.status.rank.is_any_apprentice_rank():
-        if get_clan_setting("12_moon_graduation"):
-            _ready = cat.moons >= 12
-        else:
-            _ready = (
-                cat.experience_level not in ["untrained", "learning"]
-                and cat.moons >= constants.CONFIG["graduation"]["min_graduating_age"]
-            ) or cat.moons >= constants.CONFIG["graduation"]["max_apprentice_age"][
-                cat.status.rank
-            ]
-
-        if _ready:
-            if get_clan_setting("12_moon_graduation"):
-                preparedness = "prepared"
-            else:
-                if cat.moons == constants.CONFIG["graduation"]["min_graduating_age"]:
-                    preparedness = "early"
-                elif cat.experience_level in ["untrained", "learning"]:
-                    preparedness = "unprepared"
-                else:
-                    preparedness = "prepared"
-
-            if cat.status.rank == CatRank.APPRENTICE:
-                ceremony(cat, CatRank.WARRIOR, preparedness)
-                ceremony_accessory = True
-                gain_accessories(cat)
-
-            # promote to med cat
-            elif cat.status.rank == CatRank.MEDICINE_APPRENTICE:
-                ceremony(cat, CatRank.MEDICINE_CAT, preparedness)
-                ceremony_accessory = True
-                gain_accessories(cat)
-
-            elif cat.status.rank == CatRank.MEDIATOR_APPRENTICE:
-                ceremony(cat, CatRank.MEDIATOR, preparedness)
-                ceremony_accessory = True
-                gain_accessories(cat)
-
-
-def _is_suitable_medcat_app(cat) -> bool:
-    """
-    Determines whether this cat will become a medicine cat
-    :param cat: A kitten preparing for apprenticeship ceremony
-    :return: True if the kitten should be a medcat, False otherwise
-    """
-    # assign chance to become med app depending on current med cat and traits
-    chance = constants.CONFIG["roles"]["base_medicine_app_chance"]  # 41
-    logger.info("Medcat app %s starting chance: %d", str(cat.name), chance)
-
-    med_cat_list = [
-        i
-        for i in Cat.all_cats_list
-        if i.status.rank.is_any_medicine_rank() and i.status.alive_in_player_clan
-    ]
-
-    num_medcats = len(med_cat_list)
-
-    # get number of medcat apps
-    num_med_apps = len(
-        [cat.status.rank == CatRank.MEDICINE_APPRENTICE for cat in med_cat_list]
-    )
-    logger.debug("Current number of medcats: %d", num_medcats - num_med_apps)
-    logger.debug("Current number of medcat apps: %d", num_med_apps)
-
-    # check if the Clan has sufficient med cats
-    enough_working_meds = medicine_cats_can_cover_clan(
-        Cat.all_cats.values(),
-        amount_per_med=get_amount_cat_for_one_medic(game.clan),
-    )
-
-    if (
-        floor(num_med_apps / max(1, (len(med_cat_list) - num_med_apps)))
-        > constants.CONFIG["roles"]["medicine cat apprentice"]["max_medcats_to_apps"]
-    ):
-        if enough_working_meds:
-            # early return if the ratio of apps would be too high
-            logger.info("Too many apprentices for medcat population. Aborting.")
-            return False
-        logger.debug(
-            "Too many apprentices for medcat population, but not enough medicine cats for Clan! Continuing."
-        )
-
-    # check if the medicine cats are old
-    senior_meds = [
-        c
-        for c in med_cat_list
-        if c.age == "senior" and c.status.rank == CatRank.MEDICINE_CAT
-    ]
-
-    ancient_meds = [
-        c
-        for c in senior_meds
-        if c.moons
-        >= constants.CONFIG["roles"]["medicine cat apprentice"][
-            "threshold_moons_ancient"
-        ]
-    ]
-
-    senior_med_ratio = (len(senior_meds) / num_medcats) if num_medcats != 0 else 0
-
-    ancient_med_ratio = (len(ancient_meds) / num_medcats) if num_medcats != 0 else 0
-
-    if (
-        ancient_med_ratio
-        > constants.CONFIG["roles"]["medicine cat apprentice"][
-            "threshold_percentage_ancient"
-        ]
-        / 100
-    ):
-        # These chances apply if enough medicine cats are very old.
-        if enough_working_meds:
-            chance = chance / 3
-        else:
-            logger.info("Not enough healthy medicine cats")
-            chance = chance / 14
-
-        logger.info("Ancient medicine cats, chance updated to %d", round(chance))
-    elif (
-        senior_med_ratio
-        > constants.CONFIG["roles"]["medicine cat apprentice"][
-            "threshold_percentage_seniors"
-        ]
-        / 100
-    ):
-        # These chances apply if enough medicine cats are elders.
-        if enough_working_meds:
-            chance = chance / 2.22
-        else:
-            logger.info("Not enough healthy medicine cats")
-            chance = chance / 14
-
-        logger.info("Senior medicine cats, chance updated to %d", round(chance))
-    else:
-        # These chances will only be reached if the
-        # Clan has at least one non-elder medicine cat.
-        if not enough_working_meds:
-            chance = chance / 7.125
-            logger.info(
-                "Not enough healthy medicine cats, chance updated to %d", chance
-            )
-        else:
-            chance = chance * 2.22
-            logger.info(
-                "Enough healthy young medicine cats, chance updated to %d", chance
-            )
-
-    if cat.personality.trait in [
-        "careful",
-        "compassionate",
-        "loving",
-        "wise",
-        "faithful",
-    ]:
-        chance = chance / 1.3
-        logger.info("Suitable trait, chance updated to %d", round(chance))
-
-    elif cat.personality.trait in [
-        "adventurous",
-        "arrogant",
-        "bold",
-        "bloodthirsty",
-        "cold",
-        "fierce",
-        "rebellious",
-        "troublesome",
-        "vengeful",
-    ]:
-        chance = chance * 2
-        logger.info("Unsuitable trait, chance updated to %d", round(chance))
-
-    beneficial_skills = [
-        SkillPath.OMEN,
-        SkillPath.PROPHET,
-        SkillPath.HEALER,
-        SkillPath.STAR,
-        SkillPath.DREAM,
-        SkillPath.CLAIRVOYANT,
-        SkillPath.GHOST,
-        SkillPath.CAMP,
-    ]
-
-    if cat.skills.primary.path in beneficial_skills:
-        chance = chance / 2
-        logger.info("beneficial primary skill, chance updated to %d", round(chance))
-
-    if cat.skills.secondary and cat.skills.secondary.path in beneficial_skills:
-        chance = chance / 4
-        logger.info("beneficial secondary skill, chance updated to %d", round(chance))
-
-    if cat.is_disabled():
-        chance = chance / 2
-
-    if num_med_apps == 0:
-        # if there are no apprentices at all, make it slightly easier to get one
-        logger.info("No apprentices at all")
-        chance = chance / 1.8
-        logger.info("No medcat apprentices at all, chance updated to %d", chance)
-    if num_med_apps > 1:
-        # if there's already at least one medcat app, make it harder to get another
-        chance = chance * (1 + (0.2 * (num_med_apps - 1)))
-        logger.info("%d medcat apps, chance updated to %d", num_med_apps, chance)
-
-    chance = max(1, int(chance))
-
-    success = not int(random.random() * chance)
-    logger.info("%s final chance: %d | SUCCESS: %s", cat.name, chance, success)
-    return success
-
-
-def load_ceremonies():
-    """
-    TODO: DOCS
-    """
-
-    global CEREMONY_TXT, ceremony_id_by_tag, ceremony_lang
-
-    if ceremony_lang == i18n.config.get("locale"):
-        return
-
-    CEREMONY_TXT = load_lang_resource("events/ceremonies/ceremony-master.json")
-
-    ceremony_id_by_tag = {}
-    # Sorting.
-    for ID in CEREMONY_TXT:
-        for tag in CEREMONY_TXT[ID][0]:
-            if tag in ceremony_id_by_tag:
-                ceremony_id_by_tag[tag].add(ID)
-            else:
-                ceremony_id_by_tag[tag] = {ID}
-
-    ceremony_lang = i18n.config.get("locale")
-
-
-def ceremony(cat, promoted_to, preparedness="prepared"):
-    """
-    promote cats and add to events list
-    """
-    # ceremony = []
-
-    _ment = (
-        Cat.fetch_cat(cat.mentor) if cat.mentor else None
-    )  # Grab current mentor, if they have one, before it's removed.
-    old_name = str(cat.name)
-    cat.rank_change(promoted_to)
-    cat.rank_change_traits_skill(_ment)
-
-    involved_cats = [cat.ID]  # Clearly, the cat the ceremony is about is involved.
-
-    # Time to gather ceremonies. First, lets gather all the ceremony ID's.
-
-    # ensure the right ceremonies are loaded for the given language
-    load_ceremonies()
-
-    possible_ceremonies = set()
-    dead_mentor = None
-    mentor = None
-    previous_alive_mentor = None
-    dead_parents = []
-    living_parents = []
-    mentor_type = {
-        CatRank.MEDICINE_CAT: [CatRank.MEDICINE_CAT],
-        CatRank.WARRIOR: [
-            CatRank.WARRIOR,
-            CatRank.DEPUTY,
-            CatRank.LEADER,
-            CatRank.ELDER,
-        ],
-        CatRank.MEDIATOR: [CatRank.MEDIATOR],
-    }
-
-    try:
-        # Get all the ceremonies for the role ----------------------------------------
-        possible_ceremonies.update(ceremony_id_by_tag[promoted_to])
-
-        # Get ones for prepared status ----------------------------------------------
-        if promoted_to in (CatRank.WARRIOR, CatRank.MEDICINE_CAT, CatRank.MEDIATOR):
-            possible_ceremonies = possible_ceremonies.intersection(
-                ceremony_id_by_tag[preparedness]
-            )
-
-        # Gather ones for mentor. -----------------------------------------------------
-        tags = []
-
-        # CURRENT MENTOR TAG CHECK
-        if cat.mentor:
-            if Cat.fetch_cat(cat.mentor).status.is_leader:
-                tags.append("yes_leader_mentor")
-            else:
-                tags.append("yes_mentor")
-            mentor = Cat.fetch_cat(cat.mentor)
-        else:
-            tags.append("no_mentor")
-
-        for c in reversed(cat.former_mentor):
-            if Cat.fetch_cat(c) and Cat.fetch_cat(c).dead:
-                tags.append("dead_mentor")
-                dead_mentor = Cat.fetch_cat(c)
-                break
-
-        # Unlike dead mentors, living mentors must be VALID
-        # they must have the correct status for the role the cat
-        # is being promoted too.
-        valid_living_former_mentors = []
-        for c in cat.former_mentor:
-            if Cat.fetch_cat(c).status.alive_in_player_clan:
-                if promoted_to in mentor_type:
-                    if Cat.fetch_cat(c).status.rank in mentor_type[promoted_to]:
-                        valid_living_former_mentors.append(c)
-                else:
-                    valid_living_former_mentors.append(c)
-
-        # ALL FORMER MENTOR TAG CHECKS
-        if valid_living_former_mentors:
-            #  Living Former mentors. Grab the latest living valid mentor.
-            previous_alive_mentor = Cat.fetch_cat(valid_living_former_mentors[-1])
-            if previous_alive_mentor.status.is_leader:
-                tags.append("alive_leader_mentor")
-            else:
-                tags.append("alive_mentor")
-        else:
-            # This tag means the cat has no living, valid mentors.
-            tags.append("no_valid_previous_mentor")
-
-        # Now we add the mentor stuff:
-        temp = possible_ceremonies.intersection(ceremony_id_by_tag["general_mentor"])
-
-        for t in tags:
-            temp.update(possible_ceremonies.intersection(ceremony_id_by_tag[t]))
-
-        possible_ceremonies = temp
-
-        # Gather for parents ---------------------------------------------------------
-        for p in [cat.parent1, cat.parent2]:
-            if Cat.fetch_cat(p):
-                if Cat.fetch_cat(p).dead:
-                    dead_parents.append(Cat.fetch_cat(p))
-                # For the purposes of ceremonies, living parents
-                # who are also the leader are not counted.
-                elif (
-                    Cat.fetch_cat(p).status.alive_in_player_clan
-                    and Cat.fetch_cat(p).status.rank != CatRank.LEADER
-                ):
-                    living_parents.append(Cat.fetch_cat(p))
-
-        tags = []
-        if len(dead_parents) >= 1 and "orphaned" not in cat.backstory:
-            tags.append("dead1_parents")
-        if len(dead_parents) >= 2 and "orphaned" not in cat.backstory:
-            tags.append("dead1_parents")
-            tags.append("dead2_parents")
-
-        if len(living_parents) >= 1:
-            tags.append("alive1_parents")
-        if len(living_parents) >= 2:
-            tags.append("alive2_parents")
-
-        temp = possible_ceremonies.intersection(ceremony_id_by_tag["general_parents"])
-
-        for t in tags:
-            temp.update(possible_ceremonies.intersection(ceremony_id_by_tag[t]))
-
-        possible_ceremonies = temp
-
-        # Gather for leader ---------------------------------------------------------
-
-        tags = []
-        if game.clan.leader and game.clan.leader.status.alive_in_player_clan:
-            tags.append("yes_leader")
-        else:
-            tags.append("no_leader")
-
-        temp = possible_ceremonies.intersection(ceremony_id_by_tag["general_leader"])
-
-        for t in tags:
-            temp.update(possible_ceremonies.intersection(ceremony_id_by_tag[t]))
-
-        possible_ceremonies = temp
-
-        # Gather for backstories.json ----------------------------------------------------
-        tags = []
-        if (
-            cat.backstory
-            in BACKSTORIES["backstory_categories"]["abandoned_backstories"]
-        ):
-            tags.append("abandoned")
-        elif cat.backstory == "clanborn":
-            tags.append("clanborn")
-        elif cat.backstory in BACKSTORIES["backstory_categories"]["loner_backstories"]:
-            tags.append("loner")
-        elif (
-            cat.backstory in BACKSTORIES["backstory_categories"]["kittypet_backstories"]
-        ):
-            tags.append("kittypet")
-        elif cat.backstory in BACKSTORIES["backstory_categories"]["rogue_backstories"]:
-            tags.append("rogue")
-
-        temp = possible_ceremonies.intersection(ceremony_id_by_tag["general_backstory"])
-
-        for t in tags:
-            temp.update(possible_ceremonies.intersection(ceremony_id_by_tag[t]))
-
-        possible_ceremonies = temp
-        # Check if cat does NOT have a suffix (for the sake of loner/kittypet/rogue) ----------------
-        # this also means we could probably have more easter eggs hehe
-        tags = []
-        if not cat.name.suffix:
-            tags.append("no_suffix")
-
-        # Gather for traits --------------------------------------------------------------
-
-        temp = possible_ceremonies.intersection(ceremony_id_by_tag["all_traits"])
-
-        if cat.personality.trait in ceremony_id_by_tag:
-            temp.update(
-                possible_ceremonies.intersection(
-                    ceremony_id_by_tag[cat.personality.trait]
-                )
-            )
-
-        possible_ceremonies = temp
-    except Exception as ex:
-        traceback.print_exception(type(ex), ex, ex.__traceback__)
-        print("Issue gathering ceremony text.", str(cat.name), promoted_to)
-
-    # getting the random honor if it's needed
-    random_honor = None
-    if promoted_to in (CatRank.WARRIOR, CatRank.MEDIATOR, CatRank.MEDICINE_CAT):
-        traits = load_lang_resource("events/ceremonies/ceremony_traits.json")
-
-        try:
-            random_honor = random.choice(traits[cat.personality.trait])
-        except KeyError:
-            random_honor = i18n.t("defaults.ceremony_honor")
-
-    if cat.status.rank in (CatRank.WARRIOR, CatRank.MEDICINE_CAT, CatRank.MEDIATOR):
-        cat.history.add_app_ceremony(random_honor)
-
-    ceremony_tags, ceremony_text = CEREMONY_TXT[
-        random.choice(list(possible_ceremonies))
-    ]
-
-    # This is a bit strange, but it works. If there is
-    # only one parent involved, but more than one living
-    # or dead parent, the adjust text function will pick
-    # a random parent. However, we need to know the
-    # parent to include in the involved cats. Therefore,
-    # text adjust also returns the random parents it picked,
-    # which will be added to the involved cats if needed.
-    (
-        ceremony_text,
-        involved_living_parent,
-        involved_dead_parent,
-    ) = ceremony_text_adjust(
-        ceremony_text,
-        cat,
-        dead_mentor=dead_mentor,
-        random_honor=random_honor,
-        old_name=old_name,
-        mentor=mentor,
-        previous_alive_mentor=previous_alive_mentor,
-        living_parents=living_parents,
-        dead_parents=dead_parents,
-    )
-
-    # Gather additional involved cats
-    for tag in ceremony_tags:
-        if tag == "yes_leader":
-            involved_cats.append(game.clan.leader.ID)
-        elif tag in ["yes_mentor", "yes_leader_mentor"]:
-            involved_cats.append(cat.mentor)
-        elif tag == "dead_mentor":
-            involved_cats.append(dead_mentor.ID)
-        elif tag in ["alive_mentor", "alive_leader_mentor"]:
-            involved_cats.append(previous_alive_mentor.ID)
-        elif tag == "alive2_parents" and len(living_parents) >= 2:
-            for c in living_parents[:2]:
-                involved_cats.append(c.ID)
-        elif tag == "alive1_parents" and involved_living_parent:
-            involved_cats.append(involved_living_parent.ID)
-        elif tag == "dead2_parents" and len(dead_parents) >= 2:
-            for c in dead_parents[:2]:
-                involved_cats.append(c.ID)
-        elif tag == "dead1_parent" and involved_dead_parent:
-            involved_cats.append(involved_dead_parent.ID)
-
-    # remove duplicates
-    involved_cats = list(set(involved_cats))
-
-    game.cur_events_list.append(Single_Event(ceremony_text, "ceremony", involved_cats))
-    # game.ceremony_events_list.append(f'{cat.name}{ceremony_text}')
 
 
 def gain_accessories(cat):
@@ -2583,5 +1991,4 @@ def check_and_promote_deputy():
         game.cur_events_list.append(Single_Event(text, "ceremony", involved_cats))
 
 
-load_ceremonies()
 load_war_resources()
